POC for segment replication. (#2075)
* Initial POC for segment replication.

In this POC replicas are configured as read only by not creating an indexwriter.
After primary shards refresh, a checkpoint is sent over the transport layer to replicas.
Once received, replicas fetch files in the checkpoint from the primary shard.
This initial commit ignores failover, retention leases, and shard allocation.

Signed-off-by: Marc Handalian <handalm@amazon.com>

* Remove bypass of fsync on primaries and force fsync on replicas.

This change will force an fsync on replicas when a new commit point is received.

Signed-off-by: Marc Handalian <handalm@amazon.com>

* Fix replicas from processing checkpoints from other indices.

Signed-off-by: Marc Handalian <handalm@amazon.com>

* Remove explicit fsync when every file is copied.

Signed-off-by: Marc Handalian <handalm@amazon.com>

* Fix recovery states to move to completed during intiial shard recovery and mark the shard as active.

With this change IndexShard.startRecovery will only set up a replica shard and mark
it as tracked with the primary.  We will then only start replication after the
primary has refreshed after performing the first operation.
This also avoids a condition when the initial recovery is trying to replicate
from a primary shard that has not performed any operations and waits indefinately for
a replica to catch up to the latest sequence number.  This change also ensures that
we are only ever performing one replication event at any given moment.

Signed-off-by: Marc Handalian <handalm@amazon.com>

* Ignore replication checkpoints if we are already up to the published checkpoint.

This change ensures we do not start a replication sequence if we already have the checkpoint.
This changes the checkpoint published from the primary to the latest processed checkpoint instead of the latest persisted.

Signed-off-by: Marc Handalian <handalm@amazon.com>

* Fix retention lease invariant in ReplicationTracker.

To satisfy this invariant, This change updates the TRACK_SHARD action to clone
the primary's retention lease and use it as the replicas.

Signed-off-by: Marc Handalian <handalm@amazon.com>

* Fix SegmentReplicationPrimaryService to wait until replica ShardRouting is setup.

Fixes to sleep the thread instead of incorrectly using a monitor.

Signed-off-by: Marc Handalian <handalm@amazon.com>

* Remove duplicate method to fetch local checkpoint.

Signed-off-by: Marc Handalian <handalm@amazon.com>

* Fix startup when replicas are not listed in primary's routing table.

This removes a wait in favor of throwing a retryable exception.

Signed-off-by: Marc Handalian <handalm@amazon.com>

* PR cleanup.

Renamed SegmentReplicationService -> SegmentReplicationReplicaService.
Removed if conditions in SyncedFlushService and TransportShardFlushAction.
Improved comments and documentation.

Signed-off-by: Marc Handalian <handalm@amazon.com>